CDEC-438: Move sin binned Ssc modules back home
