+++ infra/cardano-sl-infra.cabal
@@ -128,6 +128,7 @@ library
                       , cardano-sl-core
                       , cardano-sl-crypto
                       , cardano-sl-networking
+                      , cardano-sl-ssc
                       , cardano-sl-sinbin
                       , cardano-sl-util
                       , cardano-report-server >= 0.4.5
