+++ sinbin/cardano-sl-sinbin.cabal
@@ -25,13 +25,6 @@ library
                         Pos.Sinbin.Reporting.Metrics
                         Pos.Sinbin.Reporting.Wlog
 
-                        Pos.Sinbin.Ssc.Base
-                        Pos.Sinbin.Ssc.Behavior
-                        Pos.Sinbin.Ssc.Message
-                        Pos.Sinbin.Ssc.Toss.Types
-                        Pos.Sinbin.Ssc.Types
-                        Pos.Sinbin.Ssc.VssCertData
-
                         Pos.Sinbin.Shutdown
                         Pos.Sinbin.Shutdown.Class
                         Pos.Sinbin.Shutdown.Logic
