+++ ssc/src/Pos/Ssc/Message.hs
@@ -1,5 +1,78 @@
+{-# LANGUAGE DataKinds       #-}
+{-# LANGUAGE DeriveGeneric   #-}
+{-# LANGUAGE TemplateHaskell #-}
+
+-- | Messages used for communication in SSC.
+
 module Pos.Ssc.Message
-       ( module Sinbin
+       ( MCCommitment (..)
+       , MCOpening (..)
+       , MCShares (..)
+       , MCVssCertificate (..)
+       , _MCCommitment
+       , _MCOpening
+       , _MCShares
+       , _MCVssCertificate
+       , HasSscTag (..)
+       , SscTag (..)
        ) where
 
-import           Pos.Sinbin.Ssc.Message as Sinbin
+import           Universum
+
+import           Control.Lens (makePrisms)
+import           Formatting (bprint, build, (%))
+import qualified Formatting.Buildable as Buildable
+
+import           Pos.Core (StakeholderId, VssCertificate, addressHash,
+                     getCertId)
+import           Pos.Core.Ssc (InnerSharesMap, Opening, SignedCommitment)
+import           Pos.Ssc.Toss.Types (SscTag (..))
+
+class HasSscTag a where
+    toSscTag :: a -> SscTag
+
+data MCCommitment = MCCommitment !SignedCommitment
+    deriving (Show, Eq, Generic)
+
+data MCOpening = MCOpening !StakeholderId !Opening
+    deriving (Show, Eq, Generic)
+
+data MCShares = MCShares !StakeholderId !InnerSharesMap
+    deriving (Show, Eq, Generic)
+
+data MCVssCertificate = MCVssCertificate !VssCertificate
+    deriving (Show, Eq, Generic)
+
+makePrisms ''MCCommitment
+makePrisms ''MCOpening
+makePrisms ''MCShares
+makePrisms ''MCVssCertificate
+
+
+instance Buildable MCCommitment where
+    build (MCCommitment (pk, _, _))  =
+        bprint ("commitment contents from "%build) $ addressHash pk
+
+instance Buildable MCOpening where
+    build (MCOpening k _) =
+        bprint ("opening contents from "%build) k
+
+instance Buildable MCShares where
+    build (MCShares k _) =
+        bprint ("shares contents from "%build) k
+
+instance Buildable MCVssCertificate where
+    build (MCVssCertificate c) =
+        bprint ("VSS certificate contents from "%build) $ getCertId c
+
+instance HasSscTag MCCommitment where
+    toSscTag _ = CommitmentMsg
+
+instance HasSscTag MCOpening where
+    toSscTag _ = OpeningMsg
+
+instance HasSscTag MCShares where
+    toSscTag _ = SharesMsg
+
+instance HasSscTag MCVssCertificate where
+    toSscTag _ = VssCertificateMsg
